SI-22: Replace the component for selecting a sequence, in-app (#66)
* feat: LegacyNumberGeneratorModal

Set up NumberGeneratorModal as is under new "Legacy" component, so it can be used with the select component instead if an implementor wishes. This component will not be updated or maintained, however

* feat: Invalidate sequences on generate

When generating a number, the sequences will change, so invalidate those queries

refs SI-22

* feat: Typedown NumberGeneratorModal

Set up NumberGeneratorModal to use Typedown component. This is a work in progress, and is not yet feature complete.

SI-22

* feat: Include sequences which have hit maximum value

Included footer, which can include results where the sequence has hit its maximum value
-- In this case, the user cannot generate the sequence

When maximum is hit for a given sequence, it will by default move state onto next sequence in generator

* feat: Exact code match

Added option to NumberGeneratorModal which can change the permissive search into an exact query on sequence code

SI-22

* fix: Generator information

When no generator is provided, the generator name is provided along with the sequence name/code

* chore: Remove ability to unset sequence

One cannot usefully unset the sequence once selected, so remove the iconButton which would otherwise allow the user to do this

SI-22

* test: Linting and tests

Test changes -- will reduce coverage but honestly it's fine for now. Also linting, ConfirmationModal has an overly restrictive proptype so swap to intl.formatMessage instead of FormattedMessage component